server/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.5
openai>=1.0.0
pypdf>=3.16.0
python-dotenv>=1.0.0
aiofiles>=23.1.0
loguru>=0.7.0
requests>=2.31.0
tiktoken>=0.4.0
numpy>=1.24.0
google-generativeai>=0.3.0